tools/sub/requirements.txt
openai-whisper
torch
torchaudio
tqdm
soundfile
tk
ffmpeg-python
